Use interface for public API. move to factory creation of SDK
