+++ dev-app/src/main/java/com/squareup/cash/paykit/devapp/MainActivityViewModel.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.squareup.cash.paykit.CashAppPayKit
+import com.squareup.cash.paykit.CashAppPayKitFactory
 import com.squareup.cash.paykit.CashAppPayKitListener
 import com.squareup.cash.paykit.PayKitState
 import com.squareup.cash.paykit.PayKitState.ReadyToAuthorize
@@ -26,7 +27,7 @@ class MainActivityViewModel : ViewModel(), CashAppPayKitListener {
 
   var currentRequestId: String? = null
 
-  private val payKitSdk = CashAppPayKit(sandboxClientID, useSandboxEnvironment = true)
+  private val payKitSdk = CashAppPayKitFactory.createSandbox(sandboxClientID)
 
   init {
     payKitSdk.registerForStateUpdates(this@MainActivityViewModel)
