+++ paykit/src/main/java/com/squareup/cash/paykit/CashAppPayKit.kt
@@ -1,63 +1,13 @@
 package com.squareup.cash.paykit
 
-import android.content.ActivityNotFoundException
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
-import android.util.Log
 import androidx.annotation.WorkerThread
-import com.squareup.cash.paykit.PayKitState.Approved
-import com.squareup.cash.paykit.PayKitState.Authorizing
-import com.squareup.cash.paykit.PayKitState.CreatingCustomerRequest
-import com.squareup.cash.paykit.PayKitState.Declined
-import com.squareup.cash.paykit.PayKitState.NotStarted
-import com.squareup.cash.paykit.PayKitState.PayKitException
-import com.squareup.cash.paykit.PayKitState.PollingTransactionStatus
-import com.squareup.cash.paykit.PayKitState.ReadyToAuthorize
-import com.squareup.cash.paykit.PayKitState.UpdatingCustomerRequest
 import com.squareup.cash.paykit.exceptions.PayKitIntegrationException
-import com.squareup.cash.paykit.models.common.NetworkResult.Failure
-import com.squareup.cash.paykit.models.common.NetworkResult.Success
+import com.squareup.cash.paykit.impl.CashAppPayKitImpl
 import com.squareup.cash.paykit.models.response.CustomerResponseData
 import com.squareup.cash.paykit.models.sdk.PayKitPaymentAction
-import com.squareup.cash.paykit.utils.orElse
 
-private const val BASE_URL_SANDBOX = "https://sandbox.api.cash.app/customer-request/v1/"
-private const val BASE_URL_PRODUCTION = "https://api.cash.app/customer-request/v1/"
-
-/**
- * @param clientId Client Identifier that should be provided by Cash PayKit integration.
- * @param useSandboxEnvironment Specify what development environment should be used.
- */
-class CashAppPayKit(
-  private val clientId: String,
-  private val useSandboxEnvironment: Boolean = false,
-) : PayKitLifecycleListener {
-
-  private var callbackListener: CashAppPayKitListener? = null
-
-  private var customerResponseData: CustomerResponseData? = null
-
-  private var currentState: PayKitState = NotStarted
-    set(value) {
-      field = value
-      callbackListener?.payKitStateDidChange(value)
-        .orElse {
-          logError(
-            "State changed to ${value.javaClass.simpleName}, but no listeners were notified." +
-              "Make sure that you've used `registerForStateUpdates` to receive PayKit state updates.",
-          )
-        }
-    }
-
-  init {
-    if (useSandboxEnvironment) {
-      NetworkManager.baseUrl = BASE_URL_SANDBOX
-    } else {
-      NetworkManager.baseUrl = BASE_URL_PRODUCTION
-    }
-  }
-
+interface CashAppPayKit {
   /**
    * Create customer request given a [PayKitPaymentAction].
    * Must be called from a background thread.
@@ -66,20 +16,7 @@ class CashAppPayKit(
    *                      Look at [PayKitPaymentAction] for more details.
    */
   @WorkerThread
-  fun createCustomerRequest(paymentAction: PayKitPaymentAction) {
-    enforceRegisteredStateUpdatesListener()
-    currentState = CreatingCustomerRequest
-    val networkResult = NetworkManager.createCustomerRequest(clientId, paymentAction)
-    when (networkResult) {
-      is Failure -> {
-        currentState = PayKitException(networkResult.exception)
-      }
-      is Success -> {
-        customerResponseData = networkResult.data.customerResponseData
-        currentState = ReadyToAuthorize(networkResult.data.customerResponseData)
-      }
-    }
-  }
+  fun createCustomerRequest(paymentAction: PayKitPaymentAction)
 
   /**
    * Update an existing customer request given its [requestId] an the updated definitions contained within [PayKitPaymentAction].
@@ -90,20 +27,10 @@ class CashAppPayKit(
    *                      Look at [PayKitPaymentAction] for more details.
    */
   @WorkerThread
-  fun updateCustomerRequest(requestId: String, paymentAction: PayKitPaymentAction) {
-    enforceRegisteredStateUpdatesListener()
-    currentState = UpdatingCustomerRequest
-    val networkResult = NetworkManager.updateCustomerRequest(clientId, requestId, paymentAction)
-    when (networkResult) {
-      is Failure -> {
-        currentState = PayKitException(networkResult.exception)
-      }
-      is Success -> {
-        customerResponseData = networkResult.data.customerResponseData
-        currentState = ReadyToAuthorize(networkResult.data.customerResponseData)
-      }
-    }
-  }
+  fun updateCustomerRequest(
+    requestId: String,
+    paymentAction: PayKitPaymentAction
+  )
 
   /**
    * Authorize a customer request. This function must be called AFTER `createCustomerRequest`.
@@ -112,16 +39,7 @@ class CashAppPayKit(
    * @param context Android context class.
    */
   @Throws(IllegalArgumentException::class, PayKitIntegrationException::class)
-  fun authorizeCustomerRequest(context: Context) {
-    val customerData = customerResponseData
-
-    if (customerData == null) {
-      logAndSoftCrash(PayKitIntegrationException("Can't call authorizeCustomerRequest user before calling `createCustomerRequest`. Alternatively provide your own customerData"))
-      return
-    }
-
-    authorizeCustomerRequest(context, customerData)
-  }
+  fun authorizeCustomerRequest(context: Context)
 
   /**
    * Authorize a customer request with a previously created `customerData`.
@@ -132,123 +50,37 @@ class CashAppPayKit(
   fun authorizeCustomerRequest(
     context: Context,
     customerData: CustomerResponseData,
-  ) {
-    enforceRegisteredStateUpdatesListener()
-
-    if (customerData.authFlowTriggers?.mobileUrl.isNullOrEmpty()) {
-      throw IllegalArgumentException("customerData is missing redirect url")
-    }
-    // Open Mobile URL provided by backend response.
-    val intent = Intent(Intent.ACTION_VIEW)
-    intent.data = try {
-      Uri.parse(customerData.authFlowTriggers?.mobileUrl)
-    } catch (error: NullPointerException) {
-      throw IllegalArgumentException("Cannot parse redirect url")
-    }
-
-    // Replace internal state.
-    customerResponseData = customerData
-
-    // Register for process lifecycle updates.
-    PayKitLifecycleObserver.register(this)
-
-    try {
-      context.startActivity(intent)
-    } catch (activityNotFoundException: ActivityNotFoundException) {
-      throw RuntimeException("unable to open mobileUrl")
-    }
-    currentState = Authorizing
-  }
+  )
 
   /**
    *  Register a [CashAppPayKitListener] to receive PayKit callbacks.
    */
-  fun registerForStateUpdates(listener: CashAppPayKitListener) {
-    callbackListener = listener
-  }
+  fun registerForStateUpdates(listener: CashAppPayKitListener)
 
   /**
    *  Unregister any previously registered [CashAppPayKitListener] from PayKit updates.
    */
-  fun unregisterFromStateUpdates() {
-    callbackListener = null
-    PayKitLifecycleObserver.unregister(this)
-  }
+  fun unregisterFromStateUpdates()
+}
 
-  private fun enforceRegisteredStateUpdatesListener() {
-    if (callbackListener == null) {
-      logAndSoftCrash(PayKitIntegrationException("Shouldn't call this function before registering for state updates via `registerForStateUpdates`."))
-    }
-  }
+object CashAppPayKitFactory {
 
-  private fun poolTransactionStatus() {
-    Thread {
-      val networkResult = NetworkManager.retrieveUpdatedRequestData(
-        clientId,
-        customerResponseData!!.id,
-      )
-      if (networkResult is Failure) {
-        currentState = PayKitException(networkResult.exception)
-        return@Thread
-      }
-      customerResponseData = (networkResult as Success).data.customerResponseData
-
-      if (customerResponseData?.status == "APPROVED") {
-        // Successful transaction.
-        setStateFinished(true)
-      } else {
-        // If status is pending, schedule to check again.
-        if (customerResponseData?.status == "PENDING") {
-          // TODO: Add backoff strategy for long polling. ( https://www.notion.so/cashappcash/Implement-Long-pooling-retry-logic-a9af47e2db9242faa5d64df2596fd78e )
-          Thread.sleep(500)
-          poolTransactionStatus()
-          return@Thread
-        }
-
-        // Unsuccessful transaction.
-        setStateFinished(false)
-      }
-    }.start()
-  }
-
-  private fun logError(errorMessage: String) {
-    Log.e("PayKit", errorMessage)
+  /**
+   * @param clientId Client Identifier that should be provided by Cash PayKit integration.
+   */
+  fun create(
+    clientId: String,
+  ): CashAppPayKit {
+    return CashAppPayKitImpl(clientId, false)
   }
 
   /**
-   * This function will log in production, additionally it will throw an exception in sandbox or debug mode.
+   * @param clientId Client Identifier that should be provided by Cash PayKit integration.
    */
-  @Throws
-  private fun logAndSoftCrash(exception: Exception) {
-    logError("Error occurred. E.: $exception")
-    if (useSandboxEnvironment || BuildConfig.DEBUG) {
-      throw exception
-    }
-  }
-
-  private fun setStateFinished(wasSuccessful: Boolean) {
-    PayKitLifecycleObserver.unregister(this)
-    currentState = if (wasSuccessful) {
-      Approved(customerResponseData!!)
-    } else {
-      Declined
-    }
-  }
-
-  /**
-   * Lifecycle callbacks.
-   */
-
-  override fun onApplicationForegrounded() {
-    logError("onApplicationForegrounded")
-    if (currentState is Authorizing) {
-      currentState = PollingTransactionStatus
-      poolTransactionStatus()
-    }
-  }
-
-  override fun onApplicationBackgrounded() {
-    logError("onApplicationBackgrounded")
+  fun createSandbox(
+    clientId: String,
+  ): CashAppPayKit {
+    return CashAppPayKitImpl(clientId, true)
   }
 }
