+++ paykit/src/test/java/com/squareup/cash/paykit/CashAppPayKitAuthorizeTests.kt
@@ -2,6 +2,7 @@ package com.squareup.cash.paykit
 
 import android.content.Context
 import com.squareup.cash.paykit.exceptions.PayKitIntegrationException
+import com.squareup.cash.paykit.impl.CashAppPayKitImpl
 import com.squareup.cash.paykit.models.response.CustomerResponseData
 import io.mockk.MockKAnnotations
 import io.mockk.every
@@ -28,21 +29,21 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = PayKitIntegrationException::class)
   fun `should throw if calling authorize before createCustomer`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     payKit.registerForStateUpdates(mockk())
     payKit.authorizeCustomerRequest(context)
   }
 
   @Test(expected = PayKitIntegrationException::class)
   fun `should throw on authorizeCustomerRequest if has NOT registered for state updates`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true)
     payKit.authorizeCustomerRequest(context, customerResponseData)
   }
 
   @Test(expected = IllegalArgumentException::class)
   fun `should throw if missing mobileUrl from customer data`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true)
     payKit.registerForStateUpdates(mockk())
 
@@ -51,7 +52,7 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = IllegalArgumentException::class)
   fun `should throw if unable to parse mobile url in customer data`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true) {
       every { authFlowTriggers } returns null
     }
@@ -62,7 +63,7 @@ class CashAppPayKitAuthorizeTests {
 
   @Test(expected = RuntimeException::class)
   fun `should throw on if unable to start mobileUrl activity`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     val customerResponseData = mockk<CustomerResponseData>(relaxed = true) {
       every { authFlowTriggers } returns mockk {
         every { mobileUrl } returns "http://url"
