+++ paykit/src/test/java/com/squareup/cash/paykit/CashAppPayKitExceptionsTests.kt
@@ -1,6 +1,7 @@
 package com.squareup.cash.paykit
 
 import com.squareup.cash.paykit.exceptions.PayKitIntegrationException
+import com.squareup.cash.paykit.impl.CashAppPayKitImpl
 import io.mockk.mockk
 import org.junit.Test
 
@@ -8,13 +9,13 @@ class CashAppPayKitExceptionsTests {
 
   @Test(expected = PayKitIntegrationException::class)
   fun `should throw on createCustomerRequest if has NOT registered for state updates`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = true)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = true)
     payKit.createCustomerRequest(FakeData.oneTimePayment)
   }
 
   @Test
   fun `logAndSoftCrash should NOT crash in prod`() {
-    val payKit = CashAppPayKit(FakeData.CLIENT_ID, useSandboxEnvironment = false)
+    val payKit = CashAppPayKitImpl(FakeData.CLIENT_ID, useSandboxEnvironment = false)
     val listener = mockk<CashAppPayKitListener>(relaxed = true)
     payKit.registerForStateUpdates(listener)
     payKit.createCustomerRequest(FakeData.oneTimePayment)
